# verification/curator-service/api/openapi/openapi.yaml
openapi: 3.0.3
info:
    title: Global.health curator service API
    description: Server for CRUD operations on Global.health data
    termsOfService: 'https://global.health/terms-of-use/'
    contact:
        name: Global.health
        url: https://global.health
        email: info@global.health
    license:
        name: MIT
        url: 'https://opensource.org/licenses/MIT'
    version: 1.0.0
...
paths:
...
    /users/{id}:
        parameters:
            - name: id
              in: path
              description: The user ID
              required: true
              schema:
                  type: string
                  pattern: '^[a-f\d]{24}$'
        put:
            tags: [User]
            summary: Updates a specific user's roles
            operationId: updateUser
            requestBody:
                description: New roles for the user
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RoleArray'
            responses:
                '200':
                    $ref: '#/components/responses/200User'
                '400':
                    $ref: '#/components/responses/400'
                '403':
                    $ref: '#/components/responses/403'
                '404':
                    $ref: '#/components/responses/404'
                '422':
                    $ref: '#/components/responses/422'
                '500':
                    $ref: '#/components/responses/500'
        delete:
            tags: [User]
            summary: Deletes a user
            operationId: deleteUser
            responses:
                '204':
                    $ref: '#/components/responses/204'
                '401':
                    $ref: '#/components/responses/401'
                '403':
                    $ref: '#/components/responses/403'
                '404':
                    $ref: '#/components/responses/404'
                '500':
                    $ref: '#/components/responses/500'
...
    /users/agreeToDataAcknowledgement:
        post:
          tags: [ User ]
          summary: Agrees to data acknowledgement
          operationId: agreeToDataAcknowledgement
          responses:
              '200':
                  $ref: '#/components/responses/200User'
              '400':
                  $ref: '#/components/responses/400'
              '403':
                  $ref: '#/components/responses/403'
              '404':
                  $ref: '#/components/responses/404'
              '422':
                  $ref: '#/components/responses/422'
              '500':
                  $ref: '#/components/responses/500'
...
components:
    schemas:
...
        User:
            description: A user of the system (researcher, policy maker, etc)
            type: object
            properties:
                allOf:
                    $ref: '#/components/schemas/RoleArray'
                name:
                    type: string
                    description: Full name of the user if provided by the OAuth provider
                    example: John Doe
                    nullable: true
                email:
                    type: string
                    format: email
                    example: foo@bar.com
                googleID:
                    type: string
                    description: Unique identifier for the google account of the user
            required:
                - email
...
        Role:
            type: string
            description: Role of a user in the system
            enum:
                - admin
                - curator
                - "junior curator"
        RoleArray:
            type: object
            properties:
                roles:
                    type: array
                    items:
                        $ref: '#/components/schemas/Role'
...
        Error:
            description: An error response
            properties:
                message:
                    description: A description of the error that occurred
                    example: User with ID 1235abcd453 was not found
                    type: string
    responses:
...
        '200User':
            description: OK
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/User'
...
        '204':
            description: Empty
        '400':
            description: Malformed request
        '401':
            description: Unauthorized
        '403':
            description: Forbidden
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
        '404':
            description: Not found
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
        '422':
            description: Unprocessable entity
        '500':
            description: Internal server error
servers:
    - url: http://localhost:3001/api
      description: Local server
    - url: http://dev-data.covid-19.global.health/api
      description: Development server
    - url: http://data.covid-19.global.health/api
      description: Production server